Trying to organize control access by using structures

+++ Voron Poster/MainForm.cs
@@ -26,7 +26,6 @@ namespace Voron_Poster
 
         public List<TaskGui> Tasks = new List<TaskGui>();
         public TaskGui CurrTask;
-        public Scintilla CodeEditor = new Scintilla();
 
         public MainForm()
         {
@@ -46,22 +45,7 @@ namespace Voron_Poster
             //BindingFlags.SetProperty | BindingFlags.Instance | BindingFlags.NonPublic,
             //null, Tabs, new object[] { true });
 
-            CodeEditor.Dock = System.Windows.Forms.DockStyle.Fill;
-            CodeEditor.LineWrapping.VisualFlags = ScintillaNET.LineWrappingVisualFlags.End;
-            CodeEditor.Location = new System.Drawing.Point(0, 0);
-            CodeEditor.Margins.Margin1.AutoToggleMarkerNumber = 0;
-            CodeEditor.Margins.Margin1.IsClickable = true;
-            CodeEditor.Margins.Margin2.Width = 16;
-            CodeEditor.Name = "_scintilla";
-            CodeEditor.TabIndex = 0;
-            CodeEditor.ConfigurationManager.Language = "cs";
-            CodeEditor.Indentation.SmartIndentType = SmartIndent.CPP;
-            CodeEditor.Size = CodeBox.Size;
-            CodeEditor.Location = CodeBox.Location;
-            CodeEditor.Anchor = CodeBox.Anchor;
-            CodeEditor.TextChanged += CodeEditor_TextChanged;
-            CodeBox.Dispose();
-            CodeTab.Controls.Add(CodeEditor);
+
         }
 
         private void RenderHtml(string Html)
@@ -504,21 +488,21 @@ namespace Voron_Poster
 
         private void DeleteProfileButton_Click(object sender, EventArgs e)
         {
-            if (MessageBox.Show("Удалить профиль "+ProfileComboBox.Text+"?", 
+            if (MessageBox.Show("Удалить профиль " + ProfileComboBox.Text + "?",
                 "Удалить профиль?", MessageBoxButtons.YesNo) == System.Windows.Forms.DialogResult.Yes)
-            try
-            {
-                File.Delete(GetProfilePath(ProfileComboBox.Text));
-            }
-            catch (Exception Error)
-            {
-                MessageBox.Show(Error.Message);
-            }
-            finally
-            {
-                DeleteProfileButton.Enabled = false;
-                LoadProfileButton.Enabled = false;
-            }
+                try
+                {
+                    File.Delete(GetProfilePath(ProfileComboBox.Text));
+                }
+                catch (Exception Error)
+                {
+                    MessageBox.Show(Error.Message);
+                }
+                finally
+                {
+                    DeleteProfileButton.Enabled = false;
+                    LoadProfileButton.Enabled = false;
+                }
         }
 
         private void SaveProfileButton_Click(object sender, EventArgs e)
@@ -703,202 +687,204 @@ namespace Voron_Poster
 
         #endregion
 
-        string OpenedScript;
+        //public partial struct ScriptsControls
+        //{
 
-        private void TestBox_Enter(object sender, EventArgs e)
-        {
-            if ((sender as TextBox).ForeColor == SystemColors.GrayText)
-            {
-                (sender as TextBox).Text = String.Empty;
-                (sender as TextBox).ForeColor = SystemColors.WindowText;
-                (sender as TextBox).Font = new Font((sender as TextBox).Font, FontStyle.Regular);
-            }
-        }
 
-        private void TestSubjectBox_Leave(object sender, EventArgs e)
-        {
-            if (TestSubjectBox.Text == String.Empty)
-            {
-                TestSubjectBox.Text = "Тема сообщения";
-                TestSubjectBox.ForeColor = SystemColors.GrayText;
-                TestSubjectBox.Font = new Font(TestSubjectBox.Font, FontStyle.Italic);
-            }
-        }
 
-        private void TestMessageBox_Leave(object sender, EventArgs e)
-        {
-            if (TestMessageBox.Text == String.Empty)
+         partial void TestBox_Enter(object sender, EventArgs e)
             {
-                TestMessageBox.Text = "[b]Тестовое сообщение[b]\nСегодня [color=red]" +
-                    "хорошая[/color] погода.\nМы пойдем [color=#12830a]купаться[/color] на речку.";
-                TestMessageBox.ForeColor = SystemColors.GrayText;
-                TestMessageBox.Font = new Font(TestMessageBox.Font, FontStyle.Italic);
-            }
-        }
-
-        private void ScriptsPage_Enter(object sender, EventArgs e)
-        {
-            AllScriptsListBox.Items.Clear();
-            ScriptNameBox.AutoCompleteCustomSource.Clear();
-            Directory.CreateDirectory(".\\Scripts\\");
-            string[] Paths = Directory.GetFiles(".\\Scripts\\", "*.cs");
-            //trying complicated constructions =D
-            Array.ForEach<string>(Paths, s =>
-            {
-                s = s.Replace(".\\Scripts\\", String.Empty).Replace(".cs", String.Empty);
-                AllScriptsListBox.Items.Add(s);
-                ScriptNameBox.AutoCompleteCustomSource.Add(s);
-            });
-            AllScriptsListBox.SelectedIndex = AllScriptsListBox.Items.Count - 1;
-        }
-
-        private void SelectScriptInList()
-        {
-            //if (!ScriptNameBox.Text.Equals((string)AllScriptsListBox.SelectedItem,
-            //        StringComparison.OrdinalIgnoreCase) ^ SaveScriptButton.Enabled)
-            //{
-            int i;
-            //if () i = int.MaxValue;
-            //else
-            if (SaveScriptButton.Enabled) AllScriptsListBox.SelectedIndexChanged -= AllScriptsListBox_SelectedIndexChanged;
-            for (i = 0; i < AllScriptsListBox.Items.Count; i++)
-            {
-                if (ScriptNameBox.Text.Equals((string)AllScriptsListBox.Items[i],
-                    StringComparison.OrdinalIgnoreCase))
+                if ((sender as TextBox).ForeColor == SystemColors.GrayText)
                 {
-                    AllScriptsListBox.SelectedItem = AllScriptsListBox.Items[i];
-                    break;
+                    (sender as TextBox).Text = String.Empty;
+                    (sender as TextBox).ForeColor = SystemColors.WindowText;
+                    (sender as TextBox).Font = new Font((sender as TextBox).Font, FontStyle.Regular);
                 }
             }
-            if (i >= AllScriptsListBox.Items.Count)
+
+        partial void SubjectBox_Leave(object sender, EventArgs e)
             {
-                AllScriptsListBox.SelectedIndexChanged -= AllScriptsListBox_SelectedIndexChanged;
-                AllScriptsListBox.SelectedIndex = -1;
-                if (CodeEditor.Text != String.Empty)
+                if (SubjectBox.Text == String.Empty)
                 {
-                    OpenedScript = (string)AllScriptsListBox.SelectedItem;
-                    if (ScriptNameBox.Text != String.Empty)
-                        SaveScriptButton.Enabled = true;
+                    SubjectBox.Text = "Тема сообщения";
+                    SubjectBox.ForeColor = SystemColors.GrayText;
+                    SubjectBox.Font = new Font(SubjectBox.Font, FontStyle.Italic);
                 }
             }
-            AllScriptsListBox.SelectedIndexChanged += AllScriptsListBox_SelectedIndexChanged;
-            //}
 
-        }
-
-        private void ScriptNameBox_TextChanged(object sender, EventArgs e)
-        {
-            Console.WriteLine(ScriptNameBox.Text);
-            if (!ScriptNameBox.Text.Equals((string)AllScriptsListBox.SelectedItem,
-                   StringComparison.OrdinalIgnoreCase) && CodeEditor.Text != String.Empty && ScriptNameBox.Text != String.Empty)
-                SaveScriptButton.Enabled = true;
-            SelectScriptInList();
-        }
-
-        private bool AskSave()
-        {
-            switch (MessageBox.Show("Сохранить изменения?", "Изменения", MessageBoxButtons.YesNoCancel))
+        partial void MessageBox_Leave(object sender, EventArgs e)
             {
-                case System.Windows.Forms.DialogResult.Yes:
-                    SaveScriptButton_Click(SaveScriptButton, EventArgs.Empty);
-                    break;
-                case System.Windows.Forms.DialogResult.Cancel:
-                    if (OpenedScript == null)
-                    {
-                        AllScriptsListBox.SelectedIndexChanged -= AllScriptsListBox_SelectedIndexChanged;
-                        AllScriptsListBox.SelectedIndex = -1;
-                        AllScriptsListBox.SelectedIndexChanged += AllScriptsListBox_SelectedIndexChanged;
-                    }
-                    else
-                    {
-                        ScriptNameBox.Text = OpenedScript;
-                        SelectScriptInList();
-                    }
-                    return false;
+                if (MessageBox.Text == String.Empty)
+                {
+                    MessageBox.Text = "[b]Тестовое сообщение[b]\nСегодня [color=red]" +
+                        "хорошая[/color] погода.\nМы пойдем [color=#12830a]купаться[/color] на речку.";
+                    MessageBox.ForeColor = SystemColors.GrayText;
+                    MessageBox.Font = new Font(MessageBox.Font, FontStyle.Italic);
+                }
             }
-            return true;
-        }
 
-        private void AllScriptsListBox_SelectedIndexChanged(object sender, EventArgs e)
-        {
-            Console.WriteLine(AllScriptsListBox.SelectedIndex);
-            if (OpenedScript != (string)AllScriptsListBox.SelectedItem)
+        partial void ScriptsPage_Enter(object sender, EventArgs e)
             {
-                if (SaveScriptButton.Enabled && !AskSave()) return;
-                CodeEditor.Text = String.Empty;
+                ListBox.Items.Clear();
+                NameBox.AutoCompleteCustomSource.Clear();
+                Directory.CreateDirectory(".\\Scripts\\");
+                string[] Paths = Directory.GetFiles(".\\Scripts\\", "*.cs");
+                for (int i = 0; i < Paths.Length; i++)
+                {
+                    Paths[i] = Paths[i].Replace(".\\Scripts\\", String.Empty).Replace(".cs", String.Empty);
+                }
+                ListBox.Items.AddRange(Paths);
+                NameBox.AutoCompleteCustomSource.AddRange(Paths);
+                ListBox.SelectedIndex = ListBox.Items.Count - 1;
+            }
+
+        partial void SelectScriptInList()
+            {
+                //if (!NameBox.Text.Equals((string)ListBox.SelectedItem,
+                //        StringComparison.OrdinalIgnoreCase) ^ SaveButton.Enabled)
+                //{
+                int i;
+                //if () i = int.MaxValue;
+                //else
+                if (SaveButton.Enabled) ListBox.SelectedIndexChanged -= ListBox_SelectedIndexChanged;
+                for (i = 0; i < ListBox.Items.Count; i++)
+                {
+                    if (NameBox.Text.Equals((string)ListBox.Items[i],
+                        StringComparison.OrdinalIgnoreCase))
+                    {
+                        ListBox.SelectedItem = ListBox.Items[i];
+                        break;
+                    }
+                }
+                if (i >= ListBox.Items.Count)
+                {
+                    ListBox.SelectedIndexChanged -= ListBox_SelectedIndexChanged;
+                    ListBox.SelectedIndex = -1;
+                    if (CodeEditor.Text != String.Empty)
+                    {
+                        OpenedScript = (string)ListBox.SelectedItem;
+                        if (NameBox.Text != String.Empty)
+                            SaveButton.Enabled = true;
+                    }
+                }
+                ListBox.SelectedIndexChanged += ListBox_SelectedIndexChanged;
+                //}
+
+            }
+
+        partial void NameBox_TextChanged(object sender, EventArgs e)
+            {
+                Console.WriteLine(NameBox.Text);
+                if (!NameBox.Text.Equals((string)ListBox.SelectedItem,
+                       StringComparison.OrdinalIgnoreCase) && CodeEditor.Text != String.Empty && NameBox.Text != String.Empty)
+                    SaveButton.Enabled = true;
+                SelectScriptInList();
+            }
+
+            public bool AskSave()
+            {
+                switch (System.Windows.Forms.MessageBox.Show("Сохранить изменения?", "Изменения", MessageBoxButtons.YesNoCancel))
+                {
+                    case System.Windows.Forms.DialogResult.Yes:
+                        SaveButton_Click(SaveButton, EventArgs.Empty);
+                        break;
+                    case System.Windows.Forms.DialogResult.Cancel:
+                        if (OpenedScript == null)
+                        {
+                            ListBox.SelectedIndexChanged -= ListBox_SelectedIndexChanged;
+                            ListBox.SelectedIndex = -1;
+                            ListBox.SelectedIndexChanged += ListBox_SelectedIndexChanged;
+                        }
+                        else
+                        {
+                            NameBox.Text = OpenedScript;
+                            SelectScriptInList();
+                        }
+                        return false;
+                }
+                return true;
+            }
+
+            public void ListBox_SelectedIndexChanged(object sender, EventArgs e)
+            {
+                Console.WriteLine(ListBox.SelectedIndex);
+                if (OpenedScript != (string)ListBox.SelectedItem)
+                {
+                    if (SaveButton.Enabled && !AskSave()) return;
+                    CodeEditor.Text = String.Empty;
+                    try
+                    {
+                        CodeEditor.Text = System.IO.File.ReadAllText(GetScriptPath((string)ListBox.SelectedItem));
+                        OpenedScript = (string)ListBox.SelectedItem;
+                    }
+                    catch (Exception Error)
+                    {
+                        System.Windows.Forms.MessageBox.Show(Error.Message);
+                    }
+                }
+                DeleteButton.Enabled = ListBox.SelectedIndex != -1;
+                NameBox.TextChanged -= NameBox_TextChanged;
+                NameBox.Text = (string)ListBox.SelectedItem;
+                NameBox.TextChanged += NameBox_TextChanged;
+                CodeEditor.TextChanged -= CodeEditor_TextChanged;
+                SaveButton.Enabled = false;
+                CodeEditor.TextChanged += CodeEditor_TextChanged;
+            }
+
+            public void CodeEditor_TextChanged(object sender, EventArgs e)
+            {
+                SaveButton.Enabled = true;
+            }
+            public void SaveButton_Click(object sender, EventArgs e)
+            {
+                SaveButton.Enabled = false;
                 try
                 {
-                    CodeEditor.Text = System.IO.File.ReadAllText(GetScriptPath((string)AllScriptsListBox.SelectedItem));
-                    OpenedScript = (string)AllScriptsListBox.SelectedItem;
-                }
-                catch (Exception Error)
-                {
-                    MessageBox.Show(Error.Message);
-                }
-            }
-            DeleteScriptButton.Enabled = AllScriptsListBox.SelectedIndex != -1;
-            ScriptNameBox.TextChanged -= ScriptNameBox_TextChanged;
-            ScriptNameBox.Text = (string)AllScriptsListBox.SelectedItem;
-            ScriptNameBox.TextChanged += ScriptNameBox_TextChanged;
-            CodeEditor.TextChanged -= CodeEditor_TextChanged;
-            SaveScriptButton.Enabled = false;
-            CodeEditor.TextChanged += CodeEditor_TextChanged;
-        }
-
-        private void CodeEditor_TextChanged(object sender, EventArgs e)
-        {
-            SaveScriptButton.Enabled = true;
-        }
-        private void SaveScriptButton_Click(object sender, EventArgs e)
-        {
-            SaveScriptButton.Enabled = false;
-            try
-            {
-                if (ScriptNameBox.Text == String.Empty) NewScriptButton_Click(sender, e);
-                System.IO.File.WriteAllText(GetScriptPath(ScriptNameBox.Text), CodeEditor.Text);
-                OpenedScript = ScriptNameBox.Text;
-                ScriptsPage_Enter(sender, e);
-            }
-            catch (Exception Error)
-            {
-                MessageBox.Show(Error.Message);
-                SaveScriptButton.Enabled = true;
-            }
-        }
-
-        private string GetScriptPath(string Path)
-        {
-            if (!Path.EndsWith(".cs")) Path += ".cs";
-            if (Path.IndexOf('\\') < 0) Path = "Scripts\\" + Path;
-            return Path;
-        }
-
-        private void NewScriptButton_Click(object sender, EventArgs e)
-        {
-            if (SaveScriptButton.Enabled && !AskSave()) return;
-            CodeEditor.Text = String.Empty;
-            int i = 1;
-            while (File.Exists(GetScriptPath("Script #" + i.ToString()))) i++;
-            ScriptNameBox.Text = "Script #" + i.ToString();
-        }
-
-        private void DeleteScriptButton_Click(object sender, EventArgs e)
-        {
-            DeleteScriptButton.Enabled = false;
-            if (MessageBox.Show(this,"Удалить скрипт " + OpenedScript + "?",
-                "Удалить скрипт?", MessageBoxButtons.YesNo) == System.Windows.Forms.DialogResult.Yes)
-                try
-                {
-                    File.Delete(GetScriptPath(OpenedScript));
+                    if (NameBox.Text == String.Empty) NewScriptButton_Click(sender, e);
+                    System.IO.File.WriteAllText(GetScriptPath(NameBox.Text), CodeEditor.Text);
+                    OpenedScript = NameBox.Text;
                     ScriptsPage_Enter(sender, e);
-                        ScriptListBox.SelectedIndex = ScriptListBox.Items.Count - 1;
                 }
                 catch (Exception Error)
                 {
-                    DeleteScriptButton.Enabled = true;
-                    MessageBox.Show(Error.Message);
+                    System.Windows.Forms.MessageBox.Show(Error.Message);
+                    SaveButton.Enabled = true;
                 }
-        }
+            }
 
+            public string GetScriptPath(string Path)
+            {
+                if (!Path.EndsWith(".cs")) Path += ".cs";
+                if (Path.IndexOf('\\') < 0) Path = "Scripts\\" + Path;
+                return Path;
+            }
+
+            public void NewScriptButton_Click(object sender, EventArgs e)
+            {
+                if (SaveButton.Enabled && !AskSave()) return;
+                CodeEditor.Text = String.Empty;
+                int i = 1;
+                while (File.Exists(GetScriptPath("Script #" + i.ToString()))) i++;
+                NameBox.Text = "Script #" + i.ToString();
+            }
+
+            public void DeleteButton_Click(object sender, EventArgs e)
+            {
+                DeleteButton.Enabled = false;
+                if (System.Windows.Forms.MessageBox.Show("Удалить скрипт " + OpenedScript + "?",
+                    "Удалить скрипт?", MessageBoxButtons.YesNo) == System.Windows.Forms.DialogResult.Yes)
+                    try
+                    {
+                        File.Delete(GetScriptPath(OpenedScript));
+                        ScriptsPage_Enter(sender, e);
+                        ListBox.SelectedIndex = ListBox.Items.Count - 1;
+                    }
+                    catch (Exception Error)
+                    {
+                        DeleteButton.Enabled = true;
+                        System.Windows.Forms.MessageBox.Show(Error.Message);
+                    }
+            }
+        //}
     }
 }
